20231227 Revision

+++ src/main/java/es/melit/melitspringbootinmobiliaria/bussiness/InmuebleService.java
@@ -8,8 +8,11 @@ import java.util.Optional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import es.melit.melitspringbootinmobiliaria.dto.InmuebleUpdateDto;
 import es.melit.melitspringbootinmobiliaria.entities.Inmueble;
 import es.melit.melitspringbootinmobiliaria.entities.InmuebleEstadoDto;
+import es.melit.melitspringbootinmobiliaria.iDao.ClienteDao;
+import es.melit.melitspringbootinmobiliaria.iDao.EmpleadoDao;
 import es.melit.melitspringbootinmobiliaria.iDao.InmuebleDao;
 import jakarta.transaction.Transactional;
 
@@ -18,10 +21,14 @@ import jakarta.transaction.Transactional;
 public class InmuebleService implements PlantillaServicio<Inmueble> {
 	
 	public InmuebleDao iDao;	
+	public ClienteDao cDao;
+	public EmpleadoDao eDao;
 	
 	@Autowired
-	public InmuebleService(InmuebleDao iDao) {
+	public InmuebleService(InmuebleDao iDao, ClienteDao cDao, EmpleadoDao eDao) {
 		this.iDao = iDao;
+		this.cDao = cDao;
+		this.eDao = eDao;
 	}
 	
 	public List<Inmueble> listado(){		
@@ -60,7 +67,7 @@ public class InmuebleService implements PlantillaServicio<Inmueble> {
 	public List<Inmueble> findByParametros (String localidad, String tipoVivienda, Integer numHabitaciones){
 		
 		 try {
-		        return iDao.findByParametros(localidad, tipoVivienda, numHabitaciones);
+		        return iDao.findByParametros(numHabitaciones, localidad, tipoVivienda);
 		    } catch (Exception e) {
 		        System.out.println(e.getMessage());
 		        throw new RuntimeException("Error inesperado en el servidor");
@@ -70,7 +77,7 @@ public class InmuebleService implements PlantillaServicio<Inmueble> {
 	public List<Inmueble> findDemandaInmueble (String localidad, String tipoVivienda, Integer numHabitaciones){
 		
 		 try {
-		        return iDao.findByParametros(localidad, tipoVivienda, numHabitaciones);
+		        return iDao.findByParametros(numHabitaciones, localidad, tipoVivienda);
 		    } catch (Exception e) {
 		        System.out.println(e.getMessage());
 		        throw new RuntimeException("Error inesperado en el servidor");
@@ -99,12 +106,18 @@ public class InmuebleService implements PlantillaServicio<Inmueble> {
 	public void guardar(Inmueble inmueble) {
 		iDao.save(inmueble);
 	}
-	
+	@Transactional
 	public void eliminar(Integer id) {
 		if(!iDao.existsById(id)) {
 			throw new IllegalStateException("Inmueble con id: " + id + " no existe");
 		}
-		iDao.deleteById(id);	
+		Inmueble in = iDao.findById(id).get(); 
+		if(in.getPublicaciones().isEmpty() && in.getTransaccion() == null) {
+			iDao.deleteById(id);
+		} else {
+			in.setActivo(false);
+		}
+			
 	}
 
 	@Transactional
@@ -142,6 +155,50 @@ public class InmuebleService implements PlantillaServicio<Inmueble> {
 		if (inmuebleActualizado.isActivo() != inmuebleActual.isActivo()) {
 	        throw new RuntimeException("Para cambiar el estado del inmueble debes usar una solicitud específica (solicitud put con url: http://localhost:8084/api/inmobiliaria/inmuebles/cambiarestado)");
 	    }
+		
+	}
+	
+	@Transactional
+	public void actualizarConId(InmuebleUpdateDto inmuebleActualizado, Integer idInmueble) {
+		Inmueble inmuebleActual = iDao.findById(idInmueble).orElseThrow(()->
+		new IllegalStateException("Inmueble con id " + idInmueble + " no existe"));
+		
+		if(inmuebleActualizado.getDescripcion() != null && inmuebleActualizado.getDescripcion().length()>0 && !Objects.equals(inmuebleActualizado.getDescripcion(), inmuebleActual.getDescripcion())) {
+			inmuebleActual.setDescripcion(inmuebleActualizado.getDescripcion());
+		}	
+		if(inmuebleActualizado.getDireccion() != null && inmuebleActualizado.getDireccion().length()>0 && !Objects.equals(inmuebleActualizado.getDireccion(), inmuebleActual.getDireccion())) {
+			inmuebleActual.setDireccion(inmuebleActualizado.getDireccion());
+		}	
+		if(inmuebleActualizado.getLocalidad() != null && inmuebleActualizado.getLocalidad().length()>0 && !Objects.equals(inmuebleActualizado.getLocalidad(), inmuebleActual.getLocalidad())) {
+			inmuebleActual.setLocalidad(inmuebleActualizado.getLocalidad());
+		}	
+		if(inmuebleActualizado.getTipoVivienda() != null && inmuebleActualizado.getTipoVivienda().length()>0 && !Objects.equals(inmuebleActualizado.getTipoVivienda(), inmuebleActual.getTipoVivienda())) {
+			inmuebleActual.setTipoVivienda(inmuebleActualizado.getTipoVivienda());
+		}
+		if(inmuebleActualizado.getNumHabitaciones() != null && inmuebleActualizado.getNumHabitaciones()!= inmuebleActual.getNumHabitaciones()) {
+			inmuebleActual.setNumHabitaciones(inmuebleActualizado.getNumHabitaciones());
+		}	
+		if(inmuebleActualizado.getPrecio() != null && inmuebleActualizado.getPrecio()!= inmuebleActual.getPrecio()) {
+			inmuebleActual.setPrecio(inmuebleActualizado.getPrecio());
+		}
+		if(inmuebleActualizado.getMetrosCuadrados() != null && inmuebleActualizado.getMetrosCuadrados()!= inmuebleActual.getMetrosCuadrados()) {
+			inmuebleActual.setMetrosCuadrados(inmuebleActualizado.getMetrosCuadrados());
+		}
+		if(inmuebleActualizado.getIdCliente() != null && inmuebleActualizado.getIdCliente() != inmuebleActual.getCliente().getIdCliente()) {
+			
+			inmuebleActual.setCliente(cDao.findById(inmuebleActualizado.getIdCliente()).get());
+		}
+		if(inmuebleActualizado.getIdEmpleado() != null && inmuebleActualizado.getIdEmpleado() != inmuebleActual.getEmpleado().getIdEmpleado()) {
+			inmuebleActual.setEmpleado(eDao.findById(inmuebleActualizado.getIdEmpleado()).get());
+		}
+		
+		if (inmuebleActualizado.isActivo() != inmuebleActual.isActivo()) {
+			inmuebleActual.setActivo(inmuebleActualizado.isActivo());
+	    }
+		
+//		if (inmuebleActualizado.isActivo() != inmuebleActual.isActivo()) {
+//	        throw new RuntimeException("Para cambiar el estado del inmueble debes usar una solicitud específica (solicitud put con url: http://localhost:8084/api/inmobiliaria/inmuebles/cambiarestado)");
+//	    }
 		
 	}
